celsius_network_transfer_function/requirements.txt
myb-celsius-network-api>=1.0.0
